feat: greatly improve code for business day range
This up simplifies the code for generating a range of business days but
replacing convoluted up and down logic with a simple nested loop.

This change also implements optional skipping of market holidays when
constructing the range.

NOTE: The returned range is a closed range.  This may change to a
half-open range if needed.