feature: provide way to see if US markets are open from any time zone.
This code will translate a given place's time to New York time then
check to see if the markets are open for trading. This takes into
account weekends and market holidays as well as time zone differences.

This is needed to determine whether or not it is possible to stream live
data.